# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это проект был разработан в рамках тестового задания от EffectiveMobile",
        "title": "YouMusic",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "khorev.valeriy@yandex.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/Neyrzx/YouMusic?tab=MIT-1-ov-file"
        },
        "version": "0.0.1"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
        "/tracks/": {
            "get": {
                "description": "List of tracks with filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracks"
                ],
                "summary": "List of tracks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit result.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset result.",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of the artist or group.",
                        "name": "artist",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Title of track.",
                        "name": "track",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "List of tracks.",
                        "name": "releasedyaer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exact link",
                        "name": "link",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.TracksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creating track",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracks"
                ],
                "summary": "Create track",
                "parameters": [
                    {
                        "description": "Create track by song and group names.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.TracksCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        },
        "/tracks/{id}/": {
            "get": {
                "description": "Retriving track",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracks"
                ],
                "summary": "Retrive track",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "track id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get track result",
                        "schema": {
                            "$ref": "#/definitions/v1.TracksRetrieveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deliting track by track id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracks"
                ],
                "summary": "Delete track",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "track id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updating the track",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracks"
                ],
                "summary": "Update the tracks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "track id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "track id",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.TrackUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        },
        "/tracks/{id}/lyric/": {
            "get": {
                "description": "Retrive lyric verse with offset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracks"
                ],
                "summary": "Retrive verse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "track id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "verse offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/v1.TrackLyricResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.TrackLyricResponse": {
            "type": "object",
            "properties": {
                "orderID": {
                    "type": "integer"
                },
                "verse": {
                    "type": "string"
                }
            }
        },
        "v1.TrackUpdateRequest": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "lyric": {
                    "type": "string"
                },
                "released": {
                    "type": "string"
                },
                "track": {
                    "type": "string"
                }
            }
        },
        "v1.TracksCreateRequest": {
            "type": "object",
            "required": [
                "group",
                "song"
            ],
            "properties": {
                "group": {
                    "type": "string",
                    "example": "Muse"
                },
                "song": {
                    "type": "string",
                    "example": "Song name"
                }
            }
        },
        "v1.TracksResponse": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "lyric": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "released": {
                    "type": "string"
                },
                "track": {
                    "type": "string"
                }
            }
        },
        "v1.TracksRetrieveResponse": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "lyric": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "released": {
                    "type": "string"
                },
                "track": {
                    "type": "string"
                }
            }
        }
    }
}